requirements.txt
# Web Framework
Flask==2.0.1
Werkzeug==2.0.1

# News Scraping
newspaper3k==0.2.8
nltk==3.6.3
requests==2.26.0
beautifulsoup4==4.9.3
lxml==4.9.3

# Configuration
python-dotenv==0.19.0

# Date/Time handling
pytz==2023.3

# HTTP Client
certifi==2023.7.22
charset-normalizer==2.0.6
urllib3==1.26.18

# Text Processing
feedparser==6.0.8
Pillow==10.0.1
